pkg/flow/internal/controller: update docs, unexport unnecessary types
